Update to clean up code for warnings

+++ src/include/multiStore_impl.h
@@ -175,7 +175,7 @@ namespace Loci {
     T *use_ptr = tmp_base_ptr  ;
     // Call placement new
     if(!std::is_trivially_default_constructible<T>::value) {
-      for(int i=0;i<allocated_sz;++i) {
+      for(size_t i=0;i<allocated_sz;++i) {
 	T * p = use_ptr + i ;
 	new(p) T() ;
       } ;
@@ -288,7 +288,7 @@ namespace Loci {
 	  tmp_base_ptr = (T *) ((uintptr_t) tmp_base_algn+(uintptr_t)STORE_ALIGN_SIZE) ;
 	T *use_ptr = tmp_base_ptr  ;
 	// Call destructor
-	for(int i=0;i<allocated_sz;++i) {
+	for(size_t i=0;i<allocated_sz;++i) {
 	  T * p = use_ptr + i ;
 	  p->~T() ;
 	}
@@ -346,7 +346,7 @@ namespace Loci {
 	  tmp_base_ptr = (T *) ((uintptr_t) tmp_base_algn+(uintptr_t)STORE_ALIGN_SIZE) ;
 	T *use_ptr = tmp_base_ptr  ;
 	// Call destructor
-	for(int i=0;i<allocated_sz;++i) {
+	for(size_t i=0;i<allocated_sz;++i) {
 	  T * p = use_ptr + i ;
 	  p->~T() ;
 	}
@@ -466,7 +466,7 @@ namespace Loci {
 	  tmp_base_ptr = (T *) ((uintptr_t) tmp_base_algn+(uintptr_t)STORE_ALIGN_SIZE) ;
 	T *use_ptr = tmp_base_ptr  ;
 	// Call destructor
-	for(int i=0;i<allocated_sz;++i) {
+	for(size_t i=0;i<allocated_sz;++i) {
 	  T * p = use_ptr + i ;
 	  p->~T() ;
 	}
@@ -529,7 +529,7 @@ namespace Loci {
 	  tmp_base_ptr = (T *) ((uintptr_t) tmp_base_algn+(uintptr_t)STORE_ALIGN_SIZE) ;
 	T *use_ptr = tmp_base_ptr  ;
 	// Call destructor
-	for(int i=0;i<allocated_sz;++i) {
+	for(size_t i=0;i<allocated_sz;++i) {
 	  T * p = use_ptr + i ;
 	  p->~T() ;
 	}
@@ -598,7 +598,7 @@ namespace Loci {
 	  tmp_base_ptr = (T *) ((uintptr_t) tmp_base_algn+(uintptr_t)STORE_ALIGN_SIZE) ;
 	T *use_ptr = tmp_base_ptr  ;
 	// Call destructor
-	for(int i=0;i<allocated_sz;++i) {
+	for(size_t i=0;i<allocated_sz;++i) {
 	  T * p = use_ptr + i ;
 	  p->~T() ;
 	}
@@ -944,7 +944,7 @@ namespace Loci {
 	  tmp_base_ptr = (T *) ((uintptr_t) tmp_base_algn+(uintptr_t)STORE_ALIGN_SIZE) ;
 	T *use_ptr = tmp_base_ptr  ;
 	// Call destructor
-	for(int i=0;i<allocated_sz;++i) {
+	for(size_t i=0;i<allocated_sz;++i) {
 	  T * p = use_ptr + i ;
 	  p->~T() ;
 	}
